+++ src/configurations/RBTConnectConfig.cpp
@@ -6,7 +6,7 @@
 
 unsigned long RBTConnectConfig::MAX_NUM_ITER    = 1e9;
 unsigned long RBTConnectConfig::MAX_NUM_STATES  = 10000;
-float RBTConnectConfig::MAX_PLANNING_TIME       = 60e3;
+int RBTConnectConfig::MAX_PLANNING_TIME         = 60e3;
 float RBTConnectConfig::D_CRIT                  = 0.03;
 float RBTConnectConfig::DELTA                   = 3.14159;
 int RBTConnectConfig::NUM_SPINES                = 7;
