+++ src/configurations/RGBMTStarConfig.cpp
@@ -6,5 +6,5 @@
 
 unsigned long RGBMTStarConfig::MAX_NUM_ITER     = 1e9;
 unsigned long RGBMTStarConfig::MAX_NUM_STATES   = 1e9;
-float RGBMTStarConfig::MAX_PLANNING_TIME        = 60e3;
+int RGBMTStarConfig::MAX_PLANNING_TIME          = 60e3;
 int RGBMTStarConfig::TERMINATE_WHEN_PATH_IS_FOUND  = false;
